Move settings types and constants to corresponding files
